requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
docker-pycreds==0.4.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
lightning-utilities==0.9.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
omegaconf==2.3.0
packaging==23.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
pytorch-lightning==2.0.5
PyYAML==6.0
requests==2.31.0
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
torch==2.0.1
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
wandb==0.15.5
yarl==1.9.2
